pyproject.toml
[project]
name = "habits-prpject"
version = "0.1.0"
description = ""
authors = [
    {name = "origin",email = "yura28049119@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "signals (>=0.0.2,<0.0.3)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.1"
django = "^5.2.5"
djangorestframework = "^3.16.1"
djangorestframework-simplejwt = "^5.5.1"
celery = "^5.5.3"
redis = "^6.4.0"
django-cors-headers = "^4.7.0"
drf-yasg = "^1.21.10"
requests = "^2.32.5"
psycopg2 = "^2.9.10"
pytest-html = "^4.1.1"
pytest-django = "^4.11.1"
pytest = "^8.4.2"
coverage = "^7.10.6"


[tool.poetry.group.link.dependencies]
isort = "^6.0.1"
mypy = "^1.17.1"
flake8 = "^7.3.0"
black = "^25.1.0"



